k8s/version: Add capability check for v1 CRD
This preparatory commit allows us to implement dual-support for v1 and
v1beta1 CRDs.

v1 CRD support was introduced in Kubernetes 1.16 [1], therefore we will
check if the apiserver is >= 1.16, and if so, set a flag signaling v1
CRD support. Once the minimum support Kubernetes version is 1.22, we can
revert this commit, as v1beta1 support for CRDs will be removed [2],
leaving only v1 CRD support. That will render this capability check
unnecessary.

[1]:
https://github.com/kubernetes/kubernetes/blob/release-1.16/CHANGELOG/CHANGELOG-1.16.md#api-changes
[2]:
https://github.com/kubernetes/kubernetes/blob/c182a599ee3f02b4b66c6d20caaa05f9a406a10f/staging/src/k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1beta1/types.go#L395

Signed-off-by: Chris Tarazi <chris@isovalent.com>