vendor: Bump controller-tools to latest
This bump brings controller-tools up-to-date with the upstream master
branch as of Sept 6, 2020, and adds support for the following:

 - Ability to request a top-level "description" field to the validation
   schema. See further below for the rationale of why this was added.
 - Ability to request a top-level "x-preserve-unknown-fields" field to
   the validation schema. This allows CNPs and CCNPs to prevent their
   `spec` or `specs` from being a null object, which K8s will prune.
   This allows us to the maintain behavior with the CNPs under v1beta1.
 - Ability to request anyOf validation. This currently doesn't have a
   use case, but was trivial enough to add.

When applying a CNP or a CCNP with a description field (see example
below), having this "description" field present in the validation schema
is required. It was previously not required, as v1beta1 CRDs permitted
"unknown" fields. Moving to v1 CRDs disallows "unknown" fields
(`preserveUnknownFields`), hence why the support was added and updated
via this commit.

Example CNP / CCNP:

```
apiVersion: "cilium.io/v2"
kind: CiliumNetworkPolicy
description: "L7 policy for getting started using Kubernetes guide"
metadata:
  name: "l7-policy"
spec:
  endpointSelector:
    matchLabels:
      id: app1
  ingress:
  - fromEndpoints:
    - matchLabels:
        id: app2
    toPorts:
    - ports:
      - port: "80"
        protocol: TCP
      rules:
        http:
        - method: "GET"
          path: "/public"
```

Signed-off-by: Chris Tarazi <chris@isovalent.com>