k8s, examples: Allow top-level desc. for CNP
This is a workaround to allow the description field to be set at the
top-level. Originally, this field was meant to be under
spec.description, but over time due to v1beta1 allowing unknown fields
to be set, the top-level description field was used. See example below
in our test manifests.

```
apiVersion: "cilium.io/v2"
kind: CiliumNetworkPolicy
description: "Policy with the same name as cnp-update-deny-ingress to test CNP update"
metadata:
  name: "cnp-update"
spec:
  endpointSelector: {}
```

Signed-off-by: Chris Tarazi <chris@isovalent.com>