k8s: Preserve unknown fields inside CNP & CCNP
To maintain old behavior (which was with v1beta1 CRDs setting
`preserveUnknownFields: true`), we did not allow a CNP to have a null
`spec` (see example below). With v1 CRDs, `preserveUnknownFields` cannot
be true, so the policy below would have its `spec` stripped, and it
would be admitted as an "empty" policy.

If CRDs rely on the preserving unknown fields,
`x-kubernetes-preserve-unknown-fields` can be set inside a property, but
must not be set globally (spec.preserveUnknownFields). This commit adds
this property to inform K8s that we'd like to preserve the `spec`
property, so that we can maintain the same behavior as having
`preserveUnknownFields: true` with v1beta1 CRDs. Note that this only
applies on the top-level fields such as `spec`, `specs`, and `status`,
and is not global.

```
apiVersion: "cilium.io/v2"
kind: CiliumNetworkPolicy
description: "Policy with the same name as cnp-update-allow-all to test CNP update"
metadata:
  name: "cnp-update"
spec:
```

Signed-off-by: Chris Tarazi <chris@isovalent.com>