bpf: Add NewMapWithOpts for creating maps
The new function is used to create maps with options specified by a
user. Currently, only one options is supported - "CheckValueSize" which
is used to enable the value size check.

We need to disable the check for the maglev inner maps, as their
mapValue's size does not match the backing array size when creating a
map.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Martynas Pumputis <m@lambda.lt>